Test whether building the same directory structure from different paths results in the same puzzlefs representation
This test has higher chances of failing when the same directory is build
from different filesystems, which happens when /tmp is a tmpfs (and
puzzlefs is not extracted in /tmp). Currently the github CI uses
ubuntu-latest, in which /tmp is not a tmpfs, so this test will not fail.

Quoting readdir:
The order in which filenames are read by successive calls to readdir()
depends on the filesystem implementation

Signed-off-by: Ariel Miculas <amiculas@cisco.com>