Implement the foreground option for the mount command
Inspired by https://github.com/cberner/fuser/issues/179#issuecomment-1072999615
The program exits when ctrl-c is pressed or when the filesystem is
unmounted (e.g. fusermount -u <mountpoint>)

Note that this issue https://github.com/zargony/fuse-rs/issues/151
prevents using the destroy callback for notifying us that the filesystem
was unmounted, so we use the drop callback as a workaround.

Signed-off-by: Ariel Miculas <amiculas@cisco.com>