Write to pipe in case mount fails, thus fixing puzzlefs hangs
For background mounts, write to the anonymous pipe in case of error and
log the error to syslog.
For foreground mounts, write to the named pipe in case of error, thus
avoiding a hang for the reader of the pipe.

Write 's' to the pipe in case of success and 'f' in case of failure.

Fixes #66

Signed-off-by: Ariel Miculas <amiculas@cisco.com>