Add support for optional compression via command line parameter
Default to no compression in puzzlefs builds and zstd compression in
tests.

This commit adds a new field in BlobRef which stores whether the blob is
compressed or not. Per blob information is useful if we want to skip
compressing the blob in cases where the compressed version has a larger
size than the uncompressed version (e.g. when the blob is already
compressed or it has a high enough entropy that it cannot be compressed
further).

Signed-off-by: Ariel Miculas <amiculas@cisco.com>