Store the list of file chunks directly instead of wrapping it in a structure
This reduces the wire size of the inode metadata because it removes a
layer of indirection. On an ~80MB ubuntu image, the metadata space
savings is ~21KB (the metadata file size dropped from 582K to 561K).

Also increment the manifest version number since this is not a backward
compatible change.

Signed-off-by: Ariel Miculas <amiculas@cisco.com>