Change daemonize to use the crates.io version instead of github
All dependencies must be hosted on crates.io, otherwise we cannot
publish the package. For now, version 0.4.1 of daemonize is good enough.
We may want to publish a fork of daemonize to crates.io in the future to
avoid getting stuck with an old dependency.
Note: since version 0.5.0, exit_action was removed from daemonize and we
rely on it for the pipe notification mechanism.

Signed-off-by: Ariel Miculas <amiculas@cisco.com>